requirements.txt
apache-beam==2.37.0
autopep8==1.6.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
cloudpickle==2.0.0
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
fastavro==1.4.10
fasteners==0.17.3
google-api-core==1.31.5
google-apitools==0.5.31
google-auth==1.35.0
google-cloud-bigquery==2.34.3
google-cloud-bigquery-storage==2.13.1
google-cloud-bigtable==1.7.1
google-cloud-core==1.7.2
google-cloud-datastore==1.15.4
google-cloud-dlp==3.6.2
google-cloud-language==1.3.1
google-cloud-pubsub==2.12.0
google-cloud-pubsublite==1.4.1
google-cloud-recommendations-ai==0.2.0
google-cloud-spanner==1.19.2
google-cloud-videointelligence==1.16.2
google-cloud-vision==1.0.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpc-google-iam-v1==0.12.4
grpcio==1.44.0
grpcio-gcp==0.2.2
grpcio-status==1.44.0
hdfs==2.7.0
httplib2==0.19.1
idna==3.3
numpy==1.21.6
oauth2client==4.1.3
orjson==3.6.7
overrides==6.1.0
packaging==21.3
pandas==1.4.2
proto-plus==1.20.3
protobuf==3.20.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pydot==1.4.2
pymongo==3.12.3
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
toml==0.10.2
typing-utils==0.1.0
typing_extensions==4.1.1
urllib3==1.26.9
